requirements.txt
# Django
Django>=5.0,<6.0

# Database
psycopg2-binary>=2.9.9  # For PostgreSQL support

# Authentication
python-decouple>=3.8  # For environment variables
django-allauth>=0.58.2  # For authentication
whitenoise>=6.6.0  # For static files in production
gunicorn>=21.2.0  # Production WSGI server

# Django Extensions
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10

# API
# djangorestframework>=3.14.0
dj-rest-auth>=4.0.0
djangorestframework-simplejwt>=5.3.1

# Development
django-debug-toolbar>=4.2.0
pylint-django>=2.5.5
pytest-django>=4.7.0
coverage>=7.3.2

# Security
python-dotenv>=1.0.0
argon2-cffi>=23.1.0

# Email
sendgrid>=6.11.0  # For sending emails in production

# File handling
Pillow>=10.1.0  # For image handling

# Internationalization
django-modeltranslation>=0.19.1

# Date and time
pytz>=2023.3.post1
python-dateutil>=2.8.2

# Utilities
requests>=2.31.0
python-slugify>=8.0.1

# Testing
factory-boy>=3.3.0
model-bakery>=1.11.0
pytest>=7.4.2
pytest-cov>=4.1.0

# Code quality
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1
django-stubs>=4.2.7
django-stubs-ext>=4.2.7
types-python-dateutil>=2.8.19.14
types-requests>=2.31.0.10
